image portrait

+++ src/main/java/io/smarthealth/organization/person/patient/service/PatientService.java
@@ -5,6 +5,8 @@
  */
 package io.smarthealth.organization.person.patient.service;
 
+import io.smarthealth.administration.config.domain.GlobalConfiguration;
+import io.smarthealth.administration.config.domain.GlobalConfigurationRepository;
 import io.smarthealth.administration.servicepoint.data.ServicePointType;
 import io.smarthealth.administration.servicepoint.domain.ServicePoint;
 import io.smarthealth.administration.servicepoint.service.ServicePointService;
@@ -19,6 +21,7 @@ import io.smarthealth.organization.facility.service.FacilityService;
 import io.smarthealth.organization.person.data.AddressData;
 import io.smarthealth.organization.person.data.ContactData;
 import io.smarthealth.organization.person.data.PersonNextOfKinData;
+import io.smarthealth.organization.person.data.PortraitData;
 import io.smarthealth.organization.person.domain.*;
 import io.smarthealth.organization.person.patient.data.PatientData;
 import io.smarthealth.organization.person.patient.data.PatientIdentifierData;
@@ -31,6 +34,7 @@ import io.smarthealth.organization.person.patient.domain.specification.PatientSp
 import io.smarthealth.sequence.SequenceNumberService;
 import io.smarthealth.sequence.Sequences;
 import java.io.*;
+import java.nio.file.Files;
 import java.sql.Connection;
 import java.sql.SQLException;
 import java.time.LocalDateTime;
@@ -46,6 +50,7 @@ import lombok.RequiredArgsConstructor;
 import net.sf.jasperreports.engine.JRException;
 import net.sf.jasperreports.engine.JasperCompileManager;
 import net.sf.jasperreports.engine.JasperReport;
+import org.apache.commons.io.FilenameUtils;
 import org.modelmapper.ModelMapper;
 import org.modelmapper.convention.MatchingStrategies;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -55,10 +60,10 @@ import org.springframework.core.io.ResourceLoader;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.domain.Specification;
+import org.springframework.http.MediaType;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.util.FileCopyUtils;
 import org.springframework.util.MultiValueMap;
 import org.springframework.web.client.RestClientException;
 import org.springframework.web.multipart.MultipartFile;
@@ -70,7 +75,7 @@ import org.springframework.web.multipart.MultipartFile;
 @Service
 @RequiredArgsConstructor
 public class PatientService {
-    
+
     private final PatientRepository patientRepository;
     private final PersonContactRepository personContactRepository;
     private final PersonAddressRepository personAddressRepository;
@@ -78,46 +83,47 @@ public class PatientService {
     private final PatientQueueRepository patientQueueRepository;
     private final VisitRepository visitRepository;
     private final PersonNextOfKinRepository personNextOfKinRepository;
-    
+
     private final ServicePointService servicePointService;
     private final FacilityService facilityService;
 //    private final VisitService visitService;
 
     private final SequenceNumberService sequenceNumberService;
-    
+
     private final ModelMapper modelMapper;
-    
+
     private final PortraitRepository portraitRepository;
-    
+
     private final PatientIdentifierService patientIdentifierService;
-    
+
     private File patientImageDirRoot;
-    
+
+    private final GlobalConfigurationRepository globalConfigurationRepository;
+
+    @Value("${upload.image.max-size:524288}")
+    Long maxSize;
+
     @Autowired
     @Qualifier("jdbcTemplate")
     private JdbcTemplate jdbcTemplate;
-    
+
     @Autowired
     private ResourceLoader resourceLoader;
-    
-    @Value("${patientimage.upload.dir}")
-    private String uploadDir;
-    
+
     public Page<Patient> fetchAllPatients(MultiValueMap<String, String> queryParams, final Pageable pageable) {
         Specification<Patient> spec = PatientSpecification.createSpecification(queryParams.getFirst("term"));
         return patientRepository.findAll(spec, pageable);
-        // return patientRepository.findAll(pageable);
     }
-    
+
     public Page<Patient> search(String keyword, Pageable page) {
         Specification<Patient> spec = PatientSpecification.createSpecification(keyword);
         return patientRepository.findAll(spec, page);
     }
-    
+
     public Patient fetchPatientByIdentityNumber(Long patientId) {
         return patientRepository.getOne(patientId);
     }
-    
+
     public Patient fetchPatientByPersonId(Long id) {
         return patientRepository.findById(id).orElseThrow(() -> APIException.notFound("Person identified by id {0} no found", id));
     }
@@ -127,7 +133,7 @@ public class PatientService {
 //        return String.valueOf("PAT" + nextPatient);
 //    }
     public Optional<PatientData> fetchPatientByPatientNumber(final String patientNumber) {
-        
+
         return patientRepository.findByPatientNumber(patientNumber)
                 .map(patientEntity -> {
                     final PatientData patient = modelMapper.map(patientEntity, PatientData.class);
@@ -140,7 +146,7 @@ public class PatientService {
                                 .collect(Collectors.toList())
                         );
                     }
-                    
+
                     final List<PersonContact> contactDetailEntities = this.personContactRepository.findByPerson(patientEntity);
                     if (contactDetailEntities != null) {
                         patient.setContact(contactDetailEntities
@@ -149,9 +155,9 @@ public class PatientService {
                                 .collect(Collectors.toList())
                         );
                     }
-                    
+
                     final List<PatientIdentifier> patientIdentifiers = this.patientIdentifierService.fetchPatientIdentifiers(patientEntity);
-                    
+
                     if (patientIdentifiers != null && !patientIdentifiers.isEmpty()) {
                         List<PatientIdentifierData> ids = new ArrayList<>();
                         for (PatientIdentifier id : patientIdentifiers) {
@@ -159,72 +165,96 @@ public class PatientService {
                         }
                         patient.setIdentifiers(ids);
                     }
-                    
+
                     patient.setFullName((patient.getGivenName() != null ? patient.getGivenName() : "") + " " + (patient.getMiddleName() != null ? patient.getMiddleName() : "").concat(" ").concat(patient.getSurname() != null ? patient.getSurname() : " "));
                     return patient;
                 });
     }
-    
-    public String deletePortrait(String patientNumber) throws IOException {
-        final Person person = findPatientOrThrow(patientNumber);
-        Portrait portrait = portraitRepository.findByPerson(person);
-        patientImageDirRoot = new File(uploadDir);
-        System.out.println("-------------Deleting patient portrait--------------");
-        File file = new File(patientImageDirRoot, portrait.getImageName());
-        if (file.exists()) {
-            if (file.delete()) {
-                System.out.println("File deleted successfully");
-            } else {
-                System.out.println("Fail to delete file");
-            }
-        }
-        this.portraitRepository.delete(portrait);
-        //delete from directory
 
-        return patientNumber;
-    }
-    
-    @Transactional
-    public Portrait createPortrait(Person person, MultipartFile file) throws IOException {
-        if (file == null) {
-            return null;
+    public void deletePortrait(String patientNumber) throws IOException {
+        final Person person = findPatientOrThrow(patientNumber);
+        Optional<Portrait> portrait = portraitRepository.findByPerson(person);
+        if (portrait.isPresent()) {
+            GlobalConfiguration config = globalConfigurationRepository.findByName("PatientPortrait").orElseThrow(() -> APIException.notFound("Patient files folder {0} not set", "PatientPortrait"));
+            patientImageDirRoot = new File(config.getValue());
+            System.out.println("-------------Deleting patient portrait--------------");
+            System.out.println("portrait.getImageName() " + portrait.get().getImageName());
+            File file = new File(patientImageDirRoot, portrait.get().getImageName());
+            if (file.exists()) {
+                if (file.delete()) {
+                    System.out.println("File deleted successfully");
+                } else {
+                    System.out.println("Fail to delete file");
+                }
+            }
+            this.portraitRepository.delete(portrait.get());
+            //delete from directory
         }
-        
+    }
+
+    @Transactional
+    public Portrait createPortrait(Patient patient, MultipartFile file) throws IOException {
+        if (file == null) {
+            throw APIException.badRequest("Image not found");
+        }
+
+        this.throwIfInvalidSize(file.getSize());
+        this.throwIfInvalidContentType(file.getContentType());
+
         File fileForPatient = patientFileOnFolder(file);
-        
+        GlobalConfiguration config = globalConfigurationRepository.findByName("PatientPortrait").orElseThrow(() -> APIException.notFound("Patient files folder {0} not set", "PatientPortrait"));
         try (
                 InputStream in = file.getInputStream();
                 OutputStream out = new FileOutputStream(fileForPatient)) {
-            FileCopyUtils.copy(in, out);
+            // FileCopyUtils.copy(in, out);
             Portrait portrait = new Portrait();
             portrait.setContentType(file.getContentType());
-            portrait.setImage(file.getBytes());
+            //portrait.setImage(file.getBytes());
             portrait.setSize(file.getSize());
-            portrait.setPerson(person);
-            portrait.setImageUrl(uploadDir);
-            portrait.setImageName(file.getOriginalFilename());
-            portraitRepository.save(portrait);
+            portrait.setPerson(patient);
+            portrait.setImageUrl(config.getValue());
+//          portrait.setImageName(file.getOriginalFilename());
+            String extension = FilenameUtils.getExtension(file.getOriginalFilename());
+            String imageName = patient.getPatientNumber().concat(".").concat(extension);
+            portrait.setImageName(imageName);
+            //delete if any existing
+            deletePortrait(patient.getPatientNumber());
+            System.out.println("file.getName() " + file.getName());
+            //String fileName = uploadService.storeFile(file, config.getValue());
+            File fileToSave = new File(config.getValue().concat(imageName));
+            Files.copy(in, fileToSave.toPath());
+
+            portrait = portraitRepository.save(portrait);
             return portrait;
         } catch (IOException ex) {
             throw new RuntimeException(ex);
         }
     }
-    
+
     @Transactional
-    public Patient createPatient(final PatientData patient) {
+    public Patient createPatient(final PatientData patient, MultipartFile file) {
         String patientNo = sequenceNumberService.next(1L, Sequences.Patient.name());
         throwifDuplicatePatientNumber(patientNo);
 
         // use strict to prevent over eager matching (happens with ID fields)
         modelMapper.getConfiguration().setMatchingStrategy(MatchingStrategies.STRICT);
-        
+
         final Patient patientEntity = modelMapper.map(patient, Patient.class);
         patientEntity.setIsAlive(true);
         patientEntity.setPatient(true);
         patientEntity.setStatus(PatientStatus.Active);
         patientEntity.setPatientNumber(patientNo);
-        
+
         final Patient savedPatient = this.patientRepository.save(patientEntity);
+        if (file != null) {
+            try {
+                createPortrait(savedPatient, file);
+            } catch (IOException ex) {
+//                Logger.getLogger(PatientService.class.getName()).log(Level.SEVERE, null, ex);
+                ex.printStackTrace();
+            }
+        }
+
         //save patients contact details
         if (patient.getContact() != null) {
             List<PersonContact> personContacts = new ArrayList<>();
@@ -234,7 +264,7 @@ public class PatientService {
                         final PersonContact contactDetailEntity = ContactData.map(contact);
                         contactDetailEntity.setPerson(savedPatient);
                         personContacts.add(contactDetailEntity);
-                        
+
                         return contactDetailEntity;
                     })
                     .collect(Collectors.toList())
@@ -296,9 +326,9 @@ public class PatientService {
             if (!patientIdentifiersList.isEmpty()) {
                 savedPatient.setIdentifications(patientIdentifiersList);
             }*/
-            
+
         }
-        
+
         Facility facilityLogged = facilityService.loggedFacility();
         if (patient.getVisitType() != null) {
             String visitid = sequenceNumberService.next(1L, Sequences.Visit.name());
@@ -319,11 +349,11 @@ public class PatientService {
                 //send to triage
                 visit.setServicePoint(servicePoint);
             } else if (patient.getVisitType().equals("EMERGENCY_VISIT")) {
-                
+
             } else if (patient.getVisitType().equals("PHARMACY_VISIT")) {
                 //send to pharmacy
                 servicePoint = servicePointService.getServicePointByType(ServicePointType.Pharmacy);
-                
+
                 visit.setServicePoint(servicePoint);
             } else if (patient.getVisitType().equals("LABORATORY_VISIT")) {
                 //send to laboratory
@@ -344,14 +374,14 @@ public class PatientService {
             queue.setStatus(true);
             queue.setUrgency(PatientQueue.QueueUrgency.Normal);
             queue.setVisit(visit);
-            
+
             patientQueueRepository.save(queue);
 //            patientQueueService.createPatientQueue(queue);
         }
-        
+
         return savedPatient;
     }
-    
+
     @Transactional
     public Patient updatePatient(String patientNumber, Patient patient) {
         try {
@@ -361,17 +391,17 @@ public class PatientService {
             throw new RestClientException("Error updating patient number" + patientNumber);
         }
     }
-    
+
     private File patientFileOnFolder(MultipartFile f) {
         return new File(this.patientImageDirRoot, f.getOriginalFilename());
     }
-    
+
     public Patient findPatientOrThrow(String patientNumber) {
         return this.patientRepository.findByPatientNumber(patientNumber)
                 .orElseThrow(() -> APIException.notFound("Patient Number {0} not found.", patientNumber));
     }
-    
-     public Optional<Patient> findByPatientNumber(String patientNumber) {
+
+    public Optional<Patient> findByPatientNumber(String patientNumber) {
         return this.patientRepository.findByPatientNumber(patientNumber);
     }
 
@@ -380,17 +410,17 @@ public class PatientService {
             throw APIException.conflict("Duplicate Patient Number {0}", patientNumber);
         }
     }
-    
+
     public boolean patientExists(String patientNumber) {
         return patientRepository.existsByPatientNumber(patientNumber);
     }
-    
+
     public PatientData convertToPatientData(Patient patient) {
         try {
             PatientData patientData = modelMapper.map(patient, PatientData.class);
             if (patient.getAddresses() != null) {
                 List<AddressData> addresses = new ArrayList<>();
-                
+
                 patient.getAddresses().forEach((address) -> {
                     AddressData addressData = modelMapper.map(address, AddressData.class);
                     addresses.add(addressData);
@@ -413,9 +443,9 @@ public class PatientService {
                 });
                 patientData.setNok(nokData);
             }
-            
+
             final List<PatientIdentifier> patientIdentifiers = this.patientIdentifierService.fetchPatientIdentifiers(patient);
-            
+
             if (patientIdentifiers != null && !patientIdentifiers.isEmpty()) {
                 List<PatientIdentifierData> ids = new ArrayList<>();
                 for (PatientIdentifier id : patientIdentifiers) {
@@ -423,10 +453,18 @@ public class PatientService {
                 }
                 patientData.setIdentifiers(ids);
             }
-            
+
+            //fetch portrait
+            Optional<Portrait> portrait = portraitRepository.findByPerson(patient);
+            if (portrait.isPresent()) {
+                PortraitData data = new PortraitData();
+                data.setImageName(portrait.get().getImageName());
+                data.setImageUrl(portrait.get().getImageUrl());
+                patientData.setPortraitData(data);
+            }
+
             patientData.setFullName((patient.getGivenName() != null ? patient.getGivenName() : "") + " " + (patient.getMiddleName() != null ? patient.getMiddleName() : "").concat(" ").concat(patient.getSurname() != null ? patient.getSurname() : " "));
             if (patient.getDateOfBirth() != null) {
-                //patientData.setAge(String.valueOf(ChronoUnit.YEARS.between(patient.getDateOfBirth(), LocalDate.now())));
                 patientData.setAge(patient.getAge());
             }
             return patientData;
@@ -435,12 +473,12 @@ public class PatientService {
             throw APIException.internalError("An error occured while converting patient data ", e.getMessage());
         }
     }
-    
+
     public void exportPatientPdfFile(HttpServletResponse response) throws SQLException, JRException, IOException {
         Connection conn = jdbcTemplate.getDataSource().getConnection();
-        
+
         InputStream path = resourceLoader.getResource("classpath:reports/patient/PatientList.jrxml").getInputStream();
-        
+
         JasperReport jasperReport = JasperCompileManager.compileReport(path);
 
         // Parameters for report
@@ -449,5 +487,19 @@ public class PatientService {
 //        JasperPrint print = JasperFillManager.fillReport(jasperReport, parameters, conn);
 //
 //        return print;
+    }
+
+    private void throwIfInvalidSize(final Long size) {
+
+        if (size > maxSize) {
+            throw APIException.badRequest("Image can''t exceed size of {0}", maxSize);
+        }
+    }
+
+    private void throwIfInvalidContentType(final String contentType) {
+        if (!contentType.contains(MediaType.IMAGE_JPEG_VALUE)
+                && !contentType.contains(MediaType.IMAGE_PNG_VALUE)) {
+            throw APIException.badRequest("Only content type {0} and {1} allowed", MediaType.IMAGE_JPEG_VALUE, MediaType.IMAGE_PNG_VALUE);
+        }
     }
 }
